+++ src/lib/api/ContactApi.js
@@ -42,3 +42,25 @@ export const contactDelete = async (token, id) => {
     },
   });
 };
+
+export const contactDetail = async (token, id) => {
+  return fetch(`${import.meta.env.VITE_URL_API}/contacts/${id}`, {
+    method: 'GET',
+    headers: {
+      Accept: 'application/json',
+      Authorization: token,
+    },
+  });
+};
+
+export const contactUpdate = async (token, id, { first_name, last_name, email, phone }) => {
+  return fetch(`${import.meta.env.VITE_URL_API}/contacts/${id}`, {
+    method: 'PUT',
+    headers: {
+      'Content-Type': 'application/json',
+      Accept: 'application/json',
+      Authorization: token,
+    },
+    body: JSON.stringify({ first_name, last_name, email, phone }),
+  });
+};